Mise à jour du projet Derniers ajout : 	- Affichage du cube avec la SDL2 (changement complet de Affichage.c, Affichage.h) Modification à faire pour obtenir une version correctes : 	- Gestion des evenement dans la boucle du main pour les operation a effectuer et les differentes interactions 	- Modifier Affichage.c afin d'afficher les differents "menus" disponibles 	- Ameliorer la separaison entre la croix et la derniere face sur Affichage.c

+++ ModificationCube.h
@@ -1,8 +1,8 @@
-extern void MelangeAleatoire(float faceB[3][3], float faceO[3][3], float faceG[3][3], float faceR[3][3], float faceY[3][3], float faceW[3][3],int nbEtapes, int nbModAlea);
+extern void MelangeAleatoire(int faceB[3][3], int faceO[3][3], int faceG[3][3], int faceR[3][3], int faceY[3][3], int faceW[3][3],int nbEtapes, int nbModAlea);
 
 
-extern void CubeManuel(float faceB[3][3], float faceO[3][3], float faceG[3][3], float faceR[3][3], float faceY[3][3], float faceW[3][3], int nbEtapes, int nbModAlea);
+extern void CubeManuel(int faceB[3][3], int faceO[3][3], int faceG[3][3], int faceR[3][3], int faceY[3][3], int faceW[3][3], int nbEtapes, int nbModAlea);
 
 
-extern void ModificationFromTxt(float faceB[3][3], float faceO[3][3], float faceG[3][3], float faceR[3][3], float faceY[3][3], float faceW[3][3]);
+extern void ModificationFromTxt(int faceB[3][3], int faceO[3][3], int faceG[3][3], int faceR[3][3], int faceY[3][3], int faceW[3][3]);
