+++ rotation.h
@@ -1,22 +1,22 @@
-extern void U(float faceB[3][3],float faceO[3][3],float faceG[3][3],float faceR[3][3],float faceY[3][3]);
-extern void Ui(float faceB[3][3],float faceO[3][3],float faceG[3][3],float faceR[3][3],float faceY[3][3]);
+extern void U(int faceB[3][3],int faceO[3][3],int faceG[3][3],int faceR[3][3],int faceY[3][3]);
+extern void Ui(int faceB[3][3],int faceO[3][3],int faceG[3][3],int faceR[3][3],int faceY[3][3]);
 
 
-extern void R(float faceB[3][3],float faceG[3][3],float faceR[3][3],float faceY[3][3],float faceW[3][3]);
-extern void Ri(float faceB[3][3],float faceG[3][3],float faceR[3][3],float faceY[3][3],float faceW[3][3]); 
+extern void R(int faceB[3][3],int faceG[3][3],int faceR[3][3],int faceY[3][3],int faceW[3][3]);
+extern void Ri(int faceB[3][3],int faceG[3][3],int faceR[3][3],int faceY[3][3],int faceW[3][3]); 
 
 
-extern void L(float faceB[3][3],float faceO[3][3],float faceG[3][3],float faceY[3][3],float faceW[3][3]);
-extern void Li(float faceB[3][3],float faceO[3][3],float faceG[3][3],float faceY[3][3],float faceW[3][3]);
+extern void L(int faceB[3][3],int faceO[3][3],int faceG[3][3],int faceY[3][3],int faceW[3][3]);
+extern void Li(int faceB[3][3],int faceO[3][3],int faceG[3][3],int faceY[3][3],int faceW[3][3]);
 
 
-extern void D(float faceB[3][3],float faceO[3][3],float faceG[3][3],float faceR[3][3],float faceW[3][3]);
-extern void Di(float faceB[3][3],float faceO[3][3],float faceG[3][3],float faceR[3][3],float faceW[3][3]);
+extern void D(int faceB[3][3],int faceO[3][3],int faceG[3][3],int faceR[3][3],int faceW[3][3]);
+extern void Di(int faceB[3][3],int faceO[3][3],int faceG[3][3],int faceR[3][3],int faceW[3][3]);
 
 
-extern void B(float faceB[3][3],float faceO[3][3],float faceG[3][3],float faceR[3][3],float faceY[3][3],float faceW[3][3]);
-extern void Bi(float faceB[3][3],float faceO[3][3],float faceG[3][3],float faceR[3][3],float faceY[3][3],float faceW[3][3]);
+extern void B(int faceB[3][3],int faceO[3][3],int faceG[3][3],int faceR[3][3],int faceY[3][3],int faceW[3][3]);
+extern void Bi(int faceB[3][3],int faceO[3][3],int faceG[3][3],int faceR[3][3],int faceY[3][3],int faceW[3][3]);
 
 
-extern void F(float faceO[3][3],float faceG[3][3],float faceR[3][3],float faceY[3][3],float faceW[3][3], float faceB[3][3]);
-extern void Fi(float faceO[3][3],float faceG[3][3],float faceR[3][3],float faceY[3][3],float faceW[3][3], float faceB[3][3]);
+extern void F(int faceO[3][3],int faceG[3][3],int faceR[3][3],int faceY[3][3],int faceW[3][3], int faceB[3][3]);
+extern void Fi(int faceO[3][3],int faceG[3][3],int faceR[3][3],int faceY[3][3],int faceW[3][3], int faceB[3][3]);
